PR feedback on code organization. removing untested code in favor of re-using tested code.
