+++ src/ClustersLayer/index.js
@@ -1,6 +1,7 @@
 import React, { memo, useCallback, useContext, useMemo, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { useSelector } from 'react-redux';
+import { featureCollection } from '@turf/helpers';
 
 import { addNewClusterMarkers, getRenderedClustersData, removeOldClusterMarkers } from './utils';
 import { CLUSTERS_MAX_ZOOM, CLUSTERS_RADIUS, LAYER_IDS, SOURCE_IDS } from '../constants';
@@ -10,14 +11,12 @@ import { getShouldEventsBeClustered, getShouldSubjectsBeClustered } from '../sel
 import { MapContext } from '../App';
 import useClusterBufferPolygon from '../hooks/useClusterBufferPolygon';
 import { useMapEventBinding, useMapLayer, useMapSource } from '../hooks';
-import { featureCollection } from '@turf/helpers';
 
 const {
-  CLUSTER_BUFFER_POLYGON_LAYER_ID,
   CLUSTERS_LAYER_ID,
 } = LAYER_IDS;
 
-const { CLUSTER_BUFFER_POLYGON_SOURCE_ID, CLUSTERS_SOURCE_ID } = SOURCE_IDS;
+const { CLUSTERS_SOURCE_ID } = SOURCE_IDS;
 
 const CLUSTER_SOURCE_CONFIG = {
   cluster: true,
@@ -31,39 +30,15 @@ const CLUSTER_LAYER_CONFIG = {
   filter: ['has', 'point_count']
 };
 
-
-const CLUSTER_BUFFER_POLYGON_LAYER_CONFIGURATION = {
-  before: CLUSTERS_LAYER_ID,
-  id: CLUSTER_BUFFER_POLYGON_LAYER_ID,
-  maxZoom: CLUSTERS_MAX_ZOOM - 1,
-  paint: {
-    'fill-color': 'rgba(60, 120, 40, 0.4)',
-    'fill-outline-color': 'rgba(20, 100, 25, 1)',
-  },
-  source: CLUSTER_BUFFER_POLYGON_SOURCE_ID,
-  type: 'fill',
-};
-const CLUSTER_BUFFER_POLYGON_SOURCE_CONFIGURATION = { type: 'geojson' };
-
 const ClustersLayer = ({ onShowClusterSelectPopup }) => {
   const map = useContext(MapContext);
 
   const clusterMarkerHashMapRef = useRef({});
 
-  const { removeClusterPolygon, renderClusterPolygon } = useClusterBufferPolygon(
-    CLUSTER_BUFFER_POLYGON_LAYER_CONFIGURATION,
-    CLUSTER_BUFFER_POLYGON_LAYER_ID,
-    CLUSTER_BUFFER_POLYGON_SOURCE_CONFIGURATION,
-    CLUSTER_BUFFER_POLYGON_SOURCE_ID
-  );
-
-  const mapImages = useSelector((state) => state.view.mapImages);
-
   const shouldEventsBeClustered = useSelector(getShouldEventsBeClustered);
   const shouldSubjectsBeClustered = useSelector(getShouldSubjectsBeClustered);
   const eventFeatureCollection = useSelector(getMapEventFeatureCollectionWithVirtualDate);
   const subjectFeatureCollection = useSelector(getMapSubjectFeatureCollectionWithVirtualPositioning);
-
   const clustersSourceData = useMemo(() => featureCollection(
     [
       ...(shouldEventsBeClustered ? eventFeatureCollection.features : []),
@@ -75,6 +50,13 @@ const ClustersLayer = ({ onShowClusterSelectPopup }) => {
     shouldSubjectsBeClustered,
     subjectFeatureCollection.features,
   ]);
+
+  useMapSource(CLUSTERS_SOURCE_ID, clustersSourceData, CLUSTER_SOURCE_CONFIG);
+  useMapLayer(CLUSTERS_LAYER_ID, 'circle', CLUSTERS_SOURCE_ID, CLUSTER_LAYER_PAINT, null, CLUSTER_LAYER_CONFIG);
+
+  const { removeClusterPolygon, renderClusterPolygon } = useClusterBufferPolygon();
+
+  const mapImages = useSelector((state) => state.view.mapImages);
 
   const updateClusterMarkersCallback = useCallback(async () => {
 
@@ -105,8 +87,6 @@ const ClustersLayer = ({ onShowClusterSelectPopup }) => {
     }
   }, [updateClusterMarkersCallback]);
 
-  useMapSource(CLUSTERS_SOURCE_ID, clustersSourceData, CLUSTER_SOURCE_CONFIG);
-  useMapLayer(CLUSTERS_LAYER_ID, 'circle', CLUSTERS_SOURCE_ID, CLUSTER_LAYER_PAINT, null, CLUSTER_LAYER_CONFIG);
   useMapEventBinding('sourcedata', onSourceData);
 
   return null;
